Simple Login LocalStorage, Upload to LocalStorage and LocalStorage to Content Added
Login:          user: admin             psw: admin
a dummy <P> tag is showing if you are logged in on the login page
forms should not show content in URL anymore(both upload and login).
only single file can be selected and a file is required to submit:
upload now adds name of file, Subject, Description and default file img. to a array of "file" objects:
Contentpage loads LocalStorage if there is a file and adds the user added files the the end of the ContentPage: